fix: change display_order INTEGER → BIGINT to hold Date.now() values
PostgreSQL INTEGER maxes at ~2.1B. Date.now() returns ~1.78T ms which
caused every upload to fail with:
  'value 1778357050327 is out of range for type integer'

Changes:
- supabase/schema.sql: display_order BIGINT (schema v3)
- supabase/migrations/001_bigint_display_order.sql: ALTER TABLE migration
- src/lib/supabase.ts: updated inline SQL doc + GalleryRow JSDoc
- src/app/admin-gallery/page.tsx: updated SetupInstructions inline SQL

+++ jasons-glass-tint/src/app/admin-gallery/page.tsx
@@ -106,7 +106,7 @@ function SetupInstructions() {
   title         text not null default '',
   categories    text[] not null default '{}',
   framing       jsonb,
-  display_order integer not null default 0,
+  display_order bigint not null default 0,
   created_at    timestamptz not null default now(),
   updated_at    timestamptz not null default now()
 );
